backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
chromadb==0.4.22
requests==2.31.0
